refactor(org): neutralize support service filenames
