+++ turbo/apps/api/src/signals/routes/zero-host.ts
@@ -10,7 +10,7 @@ import {
   getHostedSiteFiles$,
   prepareHostedSiteDeployment$,
 } from "../services/zero-host.service";
-import { rejectSuspendedOrg$ } from "../services/zero-org-suspension.service";
+import { rejectSuspendedOrg$ } from "../services/org-suspension.service";
 import { badRequestMessage, conflict, notFound } from "../../lib/error";
 import type { RouteEntry } from "../route-entry";
