+++ turbo/apps/api/src/signals/routes/zero-org-membership-requests.ts
@@ -7,7 +7,7 @@ import { bodyResultOf } from "../context/request";
 import {
   acceptMembershipRequest$,
   rejectMembershipRequest$,
-} from "../services/zero-org-membership-requests.service";
+} from "../services/org-membership-requests.service";
 import type { RouteEntry } from "../route-entry";
 
 const adminRequired = Object.freeze({
